[MAJOR] Code refactor for all except maps

+++ server/src/routers/lafItem.js
@@ -4,7 +4,13 @@ const Item = require("../models/lafModel");
 const Image = require("../models/imageModel");
 const multer = require("multer");
 const sharp = require("sharp");
-const { UNAUTHORIZED } = require("../services/http-response");
+const {
+  UNAUTHORIZED,
+  NOT_FOUND,
+  INVALID_DATA_PROVIDED,
+  INTERNAL_SERVER_ERROR,
+  OK,
+} = require("../services/http-response");
 const upload = multer({
   limits: {
     fileSize: 10000000,
@@ -37,7 +43,7 @@ router.post(
       }
     }
     const itemData = Object.keys(request.body);
-    const allowedData = ["name", "image", "location", "color"];
+    const allowedData = ["name", "image", "location", "claimedBy"];
     const isValidOperation = itemData.every((data) =>
       allowedData.includes(data)
     );
@@ -59,7 +65,6 @@ router.post(
       await item.save();
       response.status(201).send(item);
     } catch (error) {
-      console.log(error);
       response.status(500).send("Internal Server Error");
     }
   }
@@ -68,7 +73,7 @@ router.post(
 router.get("/api/item", authUser, async (request, response) => {
   try {
     const items = await Item.find({});
-    if (!items || items.length === 0) {
+    if (!items) {
       return response.status(404).send("Items not found");
     }
     response.status(200).send(items);
@@ -78,17 +83,27 @@ router.get("/api/item", authUser, async (request, response) => {
 });
 //Router to update an item with id
 router.patch("/api/item/:id", authUser, async (request, response) => {
+  if (!request.params?.id) {
+    return response.status(NOT_FOUND.status).send(NOT_FOUND.message);
+  }
   if (!request.isAdmin) {
     return response.status(UNAUTHORIZED.status).send(UNAUTHORIZED.message);
   }
   const itemData = Object.keys(request.body);
-  const allowedData = ["name", "image", "location", "color"];
+  const allowedData = ["name", "location", "claimedBy"];
   const isValidOperation = itemData.every((data) => allowedData.includes(data));
   if (!isValidOperation) {
-    return response.status(406).send("Invalid Data Provided");
+    return response
+      .status(INVALID_DATA_PROVIDED.status)
+      .send(INVALID_DATA_PROVIDED.message);
   }
   try {
-    const item = await Item.findById({ _id: request.params.id });
+    let item = null;
+    try {
+      item = await Item.findById({ _id: request.params.id });
+    } catch (e) {
+      return response.status(NOT_FOUND.status).send(NOT_FOUND.message);
+    }
     if (!item) {
       return response.status(404).send("Item not found");
     }
@@ -96,25 +111,40 @@ router.patch("/api/item/:id", authUser, async (request, response) => {
       item[data] = request.body[data];
     });
     await item.save();
-    response.status(200).send(item);
+    response.status(OK.status).send(item);
   } catch (error) {
-    response.status(500).send("Internal Server Error");
+    response
+      .status(INTERNAL_SERVER_ERROR.status)
+      .send(INTERNAL_SERVER_ERROR.message);
   }
 });
 //Router to delete an item with id
 router.delete("/api/item/:id", authUser, async (request, response) => {
+  if (!request.params?.id) {
+    return response
+      .status(INVALID_DATA_PROVIDED.status)
+      .send(INVALID_DATA_PROVIDED.message);
+  }
   if (!request.isAdmin) {
     return response.status(UNAUTHORIZED.status).send(UNAUTHORIZED.message);
   }
   try {
-    const item = await Item.findById({ _id: request.params.id });
-    if (!item) {
-      return response.status(404).send("Item not found");
+    let item = null;
+    try {
+      item = await Item.findById({ _id: request.params.id });
+    } catch (e) {
+      return response.status(NOT_FOUND.status).send(NOT_FOUND.message);
     }
+    if (!item) {
+      return response.status(NOT_FOUND.status).send(NOT_FOUND.message);
+    }
+    await Image.deleteOne({ _id: item.image });
     await Item.deleteOne({ _id: request.params.id });
-    response.status(200).send(item);
+    response.status(OK.status).send(OK.message);
   } catch (error) {
-    response.status(500).send("Internal Server Error");
+    response
+      .status(INTERNAL_SERVER_ERROR.status)
+      .send(INTERNAL_SERVER_ERROR.message);
   }
 });
 module.exports = router;
